Rename db in communes
